Add Header Section & Google Analytics
Add in the header section with my name as the name of the site since
this will be my personal portfolio site and the navigation to allow the
user to go to the different section on the page. Also, add in Google
Analytics in order to keep track of the status of the site.