Add Banner Directive & Tagline Banner
Add in the directive for the banner which will be used display the
tagline of the site, 2 random feature projects and 2 random feature
blog post because this will be use to showcase the best part of the
site when users come to the site. Also add the first banner tagline in
order to give the user a sense of what the site is all about.