Change Banner Into Carousel From Bootstrap
Change the banner into a carousel from Bootstrap because this is what
the design should be rotating banners, of features of the the site,
when the user first loads the site up. Also, change the name of the
file to carousel feature in order to know this is part of the directive
Carousel Feature.