Add Recommendations To About Section
Add recommendations to the About section because in order to give the
user easy that I am not making this up that there is some verification
that I am telling the truth. Also, move the text colour style to the
style of section about in order for all text in this section to be
white.