Update JSON And Remove My Site and Arthritis
Update the project JSON file to have a better summary and naming for
the projects that where for iOS because some of the summaries were
either missing or had grammar errors. Furthermore remove all of the
description because those are not need yet. Also remove the time since
it was not being used.

Remove both My Personal Site and Arthritis because both of those are
not actually feature project that I would like to present to the public
for now.