Add Sorting To Projects In Portfolio Section
Add in the ability to sort projects by either name or type in Portfolio
section because the projects are not sorted in the JSON file which
would make it hard to view the section.

Also, because sorting requires to use only one ngRepeat to work. So a
nested ngRepeat would not work since this would only sort the line and
not the whole portfolio. In order to fix this, remove the nested
ngRepeat and use ngClass instead, to add in CSS class row, to the last
project, in the row, which would force next project to go to the next
line.