Add Scrollspy and Scrolling Smooth To Main Page
Add in Scrollspy in order to know which section the user is currently
on and highlights the section, in the navigation. Because this will
help the user tell which section they are on. Also add in a scrolling
smooth transition effect when clicking the section, in the navigation,
since this will give the user some visual indication that the page is
moving to that section.