Add Clearfix To Every Two Projects In Portfolio
Since Webkit has an error where if a row CSS class div is inside one of
its column then it should not affect the other columns however it does
affect the other columns. So instead add in a clearfix CSS class every
two projects with the middle clearfix CSS class be used on smaller
screens with the row CSS class be outside of the ngRepeat div in order
not to be affected.