Add Graffiti And Fix Banner Sizing On All Screens
Add Graffiti font instead of a background colour for the tagline banner
because the background colour was not staying in place when the screen
size is adjusted. Furthermore, since the background image is a brick
wall a graffiti font would fit well into the background.

Fix the banner sizing issue on all screens in order for the text to not
fall below the banner because the user would not be able to read it and
it would look bad.