Remove dist Folder From .gitignore
Remove dist folder from .gitignore in order to allow the site to be
move to the master branch since this is the live for the site. Also,
since the only way of getting files onto Github is using git to push
the files to Github and thus on to the site.