Add NoScript Tag To Header
Add in NoScript tag to header because the user may not have JavaScript
enable either by propose or by corporate policy. This is just to inform
them that they need to enable JavaScript to view the site, since the
major of the site will only work using JavaScript.

Also, move the notification for IE7 or lower into the header as well
since this notification should display under the no script notification
if the user has both IE7 or lower and disable JavaScript.