Add Both Image & URL Sitemaps
Add in both an image and URL sitemaps, to the site, because this will
allow search engines to easily access the links, on the site. As well
as map out what are the images on the site. Also add in a grunt task to
easily create a sitemap, of the site, in order not having to do this
manual.