Add Collapse Details Sections For Projects
Add in a collapse details section for the project because this will
display the details section of a project to the user if they click on
the projects Icon or name. Also, add in a function to remove all spaces
and other characters form the name in order to have them be the id, of
the collapse section.