Add Communication Between Project List & Details
Add in communication to exchange the location of the project in JSON
file into project details in order to display the details of that
project. Remove the collapse details from the project list since this
is no longer being used. Also remove ‘portfolio = this’ part because it
was looked like it is not doing anything.