Fix URLs In Projects Content JSON File
Fix the URLs in the project content JSON file to have http://start the
URL because this will reduce hard coding http:// and allow for change
to https if needed. Also add in a source code for A Simple Chess Game
because it is current the only open source on the site right now.