Add View Link & Remove Team Size For Some Projects
Add in the view links to link to either to view the project site
because to show the finish product to the user or show the actually
code, of the project, because in order to show off my code to the user.

Add in a function to remove team size, from some of the projects, if
the team size is only 1 because at that point this is  not actually
team and it would look odd having to only have one person, on the team.

Fix hideProjectEndDate premature as it was not send the value into the
function.