Add Feature Project Banners Images
Add in the feature project banners images into the project because
those will be used to promote their project when the user first loads
the site. Since the user may not scroll down quickly enough to view the
projects, this will give the user a good of view some of the projects
that are on the site.