Add HTTP Service & Move Custom Services To App JS
Add in a custom service for HTTP in order for controllers to easily
access the Project JSON file and allow for easy changing of the JSON
file. Also, move the custom services to the App JavaScript file because
some controllers are in other JavaScript files and this will allow them
to access the custom services.