Fix Details Closing Instead Of Displaying Data
Fix an issue where if the project details where open first and the user
clicks on a project banner to display another project details then the
details would close instead of displaying that project details. Because
of this created a new service to better handle when to close the project
details then bootstrap.