Fix Project Details Not Going To Correct Section
When the user clicks on the project to display the details, of that
project, the screen would go to the Contact section instead of the
Portfolio section. By adding a Du Container around the Portfolio section
the links inside this section can now use Du Scroll Spy to make sure
that the section is the correct one.