Add Jekyll Blogging Engine
Add in Jekyll a blogging engine in order to have a way to start
posting blog to my site. Because this is need to show what I know
to the world.

Also, add in a CNAME so that the blog will be under blog.jstats.me
in order to show a different design from the main site and organized
how the site is laid out.