Add Function Shorten Summary On Project List
Add function to shorten the summary text on the project list in
Portfolio section because some of the summary text where too long and
would make the columns in the row harder to line up.

Also, fix some of the misspelling in the comments in order to make them
easier to read.