Add Styling For View Source Code To Align Them
Add styling in order to have the View Source Code button align in their
rows in all screen sizes because this button needs to be on all the
screen sizes since it is a very impotent button for this section.

Also, change media queue from max width 768px to max width 767px because
this is the screen size when it goes from large screen size to medium
screen size.