Fix Positioning Of View Buttons in Project Detail
Fix positioning of view buttons in the project details, on smaller size
screens, since they where in the middle of the screen. Because of this a
directive was, for the buttons, in order to have them in two areas. So
that when on a smaller size screens, the view buttons, in the project
description, will be hidden. Then the view buttons, in the sidebar, will
then be displayed. This will allow the buttons to be at the bottom of
the section always.