Update Application Code Style to Match Prettier
Update Application code style to match Prettier because as time goes
on, code styling getting to be more standardized. So that developers
do not get stuck fighting over code styling and instead focus on the
code general, thus keeping the git difference focus on what has
changed instead of the changes they did to the code styling, which
makes the git difference harder to read.