Update Packages & Add Package Lock File
Update the packages because they have not updated for a long time, and
some are no longer in use since NPM has been updating as well, so since
the package could update and break production without the developer
knowing about it.

They created a lock file so that the package version stays as is, until
the developer updates they package again on their local computer and
checks to see if these updates would break their application. Thus add
a package-lock file for this possibles.