Add resolutions to package.json to fix graceful-fs
Add `resolutions` to `package.json` to fix `graceful-fs` because they
have change how they are interacting with Node. As a result, any package
that uses `graceful-fs` to version 4 or higher. However, some packages
that use `graceful-fs` cannot be updated because they would be the
application. Yarn can go into each package and update `graceful-fs` to
the version stated in the resolution.

---
Ref:

https://stackoverflow.com/questions/55921442/how-to-fix-referenceerror-primordials-is-not-defined-in-node