Move Version 1 of Portfolio to `v1` Folder
Move the version 1 of the Portfolio application to `v1` folder because
Angular V1, Bower and Grunt are not used at all anymore. All three of
those projects are not maintained. Those need to be more out of the way
so that React and the more modern JavaScript libraries can use in the
application.