Add React From Create-React-App Package
Add React from the Create-React-App package so that there is a base to
start. Also, React is now the main JS framework that Portfolio
application is using since it is easier to maintain and lets more
developers are using it then Angular v1, which Anglaur has also switch
to a component base.