Add Link Button Component
Add the Link button component, so there is a component to have a link
style but use it as a button. However, since some functionality needs to
be a click, such as going back to a page but still having the look of a
button, furthermore, the application can follow i18n guidelines; we need
to take a button and style it like a link to get this effect.