Add Link For Project Details on Project Item Icon
Add the link for the Project Details on the Project Item's icon and let
the user click on the image to let the user get to the details from
there. For example, some users click on the image instead of the View
Details button since the user may be more visual.