Add Search Related Elements to App State
Add the search-related elements to the application state to hold the
search results and use that to display the results. As a result, remove
the selected project since that is not needed as the chosen project does
not need to be the highlight, and the user is sent to the project's
details.