Update Portfolio to Use App State Project
Update the Portfolio section to use the application state project value
so that the application uses that and limits the use of the project in
the data folder. The application should treat the project in the data
folder as if it is from a server. Thus, it makes it easier to move to a
server if needed. Plus, it helps the application's maintainability as
the number of references to the data folder reduces.