Add Select Component to Handle Dropdowns
Add the Select component to handle the dropdowns, as it can do both the
radio button and checkbox jobs without sacrificing space on the screen.
Furthermore, the Select component is mostly the same and has many
options. Thus, it should be able to do the primary functions and handle
the styling of the components.