Move Posts State to Global State
Move the posts state to the global state so that other parts of the
application can access the posts in memory since the application does
not get that much traffic, and the posts do not change much.

With the posts now in memory, it is easier to have other components go
through it instead of getting a newer copy since it will take a lot
longer and more code to make that happen.