Update Blog Component Use Children Prop For List
Update the Blog component to use children prop for the list since it can
handle the posts loading to the application, and the Post List component
can handle the display. Instead of having everything go through a Blog
and making it dependent. Thus, the Post List and Blog components are not
probable anymore, have the Blog component uses a children component so
that it can be used in other projects and can focus on the one task it
is meant to be used for.