Update Blog Component
Update the Blog component to not update after getting the posts so that
it receives the same data again since the post does not change much, so
it can still use the posts in the state. Furthermore, the traffic to the
site is low to have the post always getting the latest version.