Update Textfield.style Split Field From Error
Update the 'Textfield.style' to split the field style from the error
style, as Textarea also needs to use the field style. Unfornently the
error style needs to be duplicated since the style component does not
work with props using `css` or at the least seems unable to do so.