Fix Links to be Secure
Fix the links to be secure so that there is no leak insecure so that the
user can always access secure content and state in the secure sections
of the application. Since by staying in a secure area, the application
can keep the user's data out of the hands of bad actors.