Update Base Use `jobs` Object From Last Jobs
Update the Base use `jobs` object from the install job to see if the
following jobs can access the same values to find if we can use values
from the last job and try to build up the process before deployment.

Since getting the job variables will allow for jobs to be interconnected
and not have to duplicate the environment from the last job, which adds
too much code and complexity to the workflow.