Refactor GitHub Workflows to Use Non-Page Workflow
Refactor the GitHub Workflows to use the non-page workflow as too many
deployment hacks need to happen for the application to be displayed on
the web. Particularly. Showing secerts environment variables in the open
to be used. As a result, a new deployment workflow with the lessons
learned from the GitHub Page Workflow, where we will deploy to a host
that can use Node, React and Environment Variables without hacking a
workflow together.