Remove Fake Components From Tests
Remove the fake components from tests as if the changes affect the
components in the components. The developer can have false hope that the
change worked. Furthermore, if a child component's props change, then
the mock component needs to change as well, raising the maintenance of
the test, which can lead to developers not creating tests as they would
have a disbelief that tests are more problematic than worth it. The
developers should ensure this is not the case and thus should try to
keep the test as easy to maintain as possible.