Remove `yup-phone` Package
Remove the `yup-phone` package, as it is not working out since it is
doing false positives to valid phone numbers. Instead, the form will be
using RegEx; however, which means the user cannot use country codes
since that is hard to do for RegEx. Also, most people are in North
America, so this may be fine.