Add Sentry For Monitoring
Add Sentry for monitoring, as there may be bugs in production that the
developer must be aware of. Sentry will be able to find those bugs and
present them on the interact page with more details on the bug.
Furthermore, Sentry can also ensure the performance goes as fast as
possible.